Adjust test
