+++ api/src/main/java/com/github/skjolber/packing/api/ep/SupportPoint3D.java
@@ -1,0 +1,16 @@
+package com.github.skjolber.packing.api.ep;
+
+import java.util.List;
+
+import com.github.skjolber.packing.api.Placement3D;
+
+public class SupportPoint3D {
+
+	private Point3D point;
+
+	private int minX;
+	private int minY;
+	private int minZ;
+
+	private List<Placement3D> placements;
+}
